+++ bot/utils/recognition.py
